+++ Runtime/Attributes/ServiceAttribute.cs
@@ -6,5 +6,10 @@ namespace JakubKrizanovsky.DependencyInjection
     public sealed class ServiceAttribute : Attribute
     {
         public bool Persistent {get; set;} = false;
+        private bool? _unique = null;
+        public bool Unique {
+            get => _unique ?? Persistent;
+            set => _unique = value;
+        }
     }
 }