+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [1.0.0] - 2026-03-06
+## [1.0.0] - 2026-04-17
 ### First Release
 - Dependency Injection Framework
+
+## [1.1.0] - 2026-04-23
+### Unique Services
+- Added support for making services unique